Moved examples and updated cv7_example.c
Moved all examples into their own subdirectories within c/cpp directories
Updated the CV7 C example with more documentation and cleaned up syntax
Removed references to MSCL-embedded
