+++ .devcontainer/devcontainer.json
@@ -29,10 +29,10 @@
     "settings": {},
     "mounts": [
         "source=/dev,target=/dev,type=bind,consistency=cached",
-        "source=${localWorkspaceFolder},target=/home/microstrain/MSCL-embedded,type=bind,consistency=cached",
-        "source=${localWorkspaceFolder}/.devcontainer/.vscode-docker,target=/home/microstrain/MSCL/.vscode,type=bind,consistency=cached",
+        "source=${localWorkspaceFolder},target=/home/microstrain/mip_sdk,type=bind,consistency=cached",
+        "source=${localWorkspaceFolder}/.devcontainer/.vscode-docker,target=/home/microstrain/mip_sdk/.vscode,type=bind,consistency=cached",
         "source=/tmp/.X11-unix,target=/tmp/.X11-unix,type=bind,consistency=cached",
     ],
     "workspaceMount": "",
-    "workspaceFolder": "/home/microstrain/MSCL-embedded"
+    "workspaceFolder": "/home/microstrain/mip_sdk"
 }
