+++ examples/c/cx5_gx5_45/cx5_gx5_45_example.c
@@ -1,14 +1,14 @@
 
 /////////////////////////////////////////////////////////////////////////////
 //
-// GQ7_Example.c
+// CX5_GX5_45_Example.c
 //
-// C Example set-up program for the GQ7
+// C Example set-up program for the CX5-45 and GX5-45
 //
-// This example shows a typical setup for the GQ7 sensor in a wheeled-vehicle application using C.
-// It is not an exhaustive example of all GQ7 settings.
+// This example shows a typical setup for the GX5-45 sensor in a wheeled-vehicle application using using C.
+// It is not an exhaustive example of all GX5-45 settings.
 // If your specific setup needs are not met by this example, please consult
-// the MSCL-embedded API documentation for the proper commands.
+// the MIP SDK API documentation for the proper commands.
 //
 //
 //!@section LICENSE
@@ -27,10 +27,10 @@
 // Include Files
 ////////////////////////////////////////////////////////////////////////////////
 
-#include "example_utils.h"
+#include "../example_utils.h"
 
-#include <mip/mip_all.h>
-#include <microstrain/connections/serial/serial_port.h>
+#include <../../src/c/mip/mip_all.h>
+#include <../../src/c/microstrain/connections/serial/serial_port.h>
 #include <stdio.h>
 #include <stdlib.h>
 
@@ -42,30 +42,29 @@
 int port = -1;
 mip_interface device;
 
-//Sensor-to-vehicle frame transformation (Euler Angles)
-float sensor_to_vehicle_transformation_euler[3] = {0.0, 0.0, 0.0};
+//Sensor-to-vehicle frame rotation (Euler Angles)
+float sensor_to_vehicle_rotation_euler[3] = {0.0f, 0.0f, 0.0f};
 
-//GNSS antenna offsets
-float gnss1_antenna_offset_meters[3] = {-0.25, 0.0, 0.0};
-float gnss2_antenna_offset_meters[3] = {0.25, 0.0, 0.0};
+//GNSS antenna offset
+float gnss_antenna_offset_meters[3] = {-0.25f, 0.0f, 0.0f};
 
 //Device data stores
-mip_shared_gps_timestamp_data sensor_gps_time;
+mip_sensor_gps_timestamp_data sensor_gps_time;
 mip_sensor_scaled_accel_data  sensor_accel;
 mip_sensor_scaled_gyro_data   sensor_gyro;
 mip_sensor_scaled_mag_data    sensor_mag;
 
-mip_gnss_fix_info_data        gnss_fix_info[2];
+mip_gnss_fix_info_data        gnss_fix_info;
 
-bool gnss_fix_info_valid[2] = {false};
+bool gnss_fix_info_valid = false;
 
-mip_shared_gps_timestamp_data filter_gps_time;
+mip_filter_timestamp_data     filter_gps_time;
 mip_filter_status_data        filter_status;
 mip_filter_position_llh_data  filter_position_llh;
 mip_filter_velocity_ned_data  filter_velocity_ned;
 mip_filter_euler_angles_data  filter_euler_angles;
 
-bool filter_state_full_nav = false;
+bool filter_state_running = false;
 
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -153,38 +152,34 @@ int main(int argc, const char* argv[])
     //Note: Querying the device base rate is only one way to calculate the descriptor decimation.
     //We could have also set it directly with information from the datasheet (shown in GNSS setup).
 
-    if(mip_3dm_get_base_rate(&device, MIP_SENSOR_DATA_DESC_SET, &sensor_base_rate) != MIP_ACK_OK)
+    if(mip_3dm_imu_get_base_rate(&device, &sensor_base_rate) != MIP_ACK_OK)
          exit_gracefully("ERROR: Could not get sensor base rate format!");
 
     const uint16_t sensor_sample_rate = 100; // Hz
     const uint16_t sensor_decimation = sensor_base_rate / sensor_sample_rate;
 
     const mip_descriptor_rate sensor_descriptors[4] = {
-        { MIP_DATA_DESC_SHARED_GPS_TIME,     sensor_decimation },
-        { MIP_DATA_DESC_SENSOR_ACCEL_SCALED, sensor_decimation },
-        { MIP_DATA_DESC_SENSOR_GYRO_SCALED,  sensor_decimation },
-        { MIP_DATA_DESC_SENSOR_MAG_SCALED,   sensor_decimation },
+        { MIP_DATA_DESC_SENSOR_TIME_STAMP_GPS, sensor_decimation },
+        { MIP_DATA_DESC_SENSOR_ACCEL_SCALED,   sensor_decimation },
+        { MIP_DATA_DESC_SENSOR_GYRO_SCALED,    sensor_decimation },
+        { MIP_DATA_DESC_SENSOR_MAG_SCALED,     sensor_decimation },
     };
 
-    if(mip_3dm_write_message_format(&device, MIP_SENSOR_DATA_DESC_SET, 4, sensor_descriptors) != MIP_ACK_OK)
+    if(mip_3dm_write_imu_message_format(&device, 4, sensor_descriptors) != MIP_ACK_OK)
         exit_gracefully("ERROR: Could not set sensor message format!");
 
 
     //
-    //Setup GNSS 1 and 2 data format to 2 Hz (decimation of 1)
+    //Setup GNSS data format to 4 Hz (decimation of 1)
     //
 
     const mip_descriptor_rate gnss_descriptors[1] = {
         { MIP_DATA_DESC_GNSS_FIX_INFO, 1 }
      };
 
-    //GNSS1
-    if(mip_3dm_write_message_format(&device, MIP_GNSS1_DATA_DESC_SET, 1, gnss_descriptors) != MIP_ACK_OK)
+    //GNSS
+    if(mip_3dm_write_gps_message_format(&device, 1, gnss_descriptors) != MIP_ACK_OK)
         exit_gracefully("ERROR: Could not set GNSS1 message format!");
-
-    //GNSS2
-    if(mip_3dm_write_message_format(&device, MIP_GNSS2_DATA_DESC_SET, 1, gnss_descriptors) != MIP_ACK_OK)
-        exit_gracefully("ERROR: Could not set GNSS2 message format!");
 
 
     //
@@ -193,74 +188,54 @@ int main(int argc, const char* argv[])
 
     uint16_t filter_base_rate;
 
-    if(mip_3dm_get_base_rate(&device, MIP_FILTER_DATA_DESC_SET, &filter_base_rate) != MIP_ACK_OK)
+    if(mip_3dm_filter_get_base_rate(&device, &filter_base_rate) != MIP_ACK_OK)
          exit_gracefully("ERROR: Could not get filter base rate format!");
 
     const uint16_t filter_sample_rate = 100; // Hz
     const uint16_t filter_decimation = filter_base_rate / filter_sample_rate;
 
     const mip_descriptor_rate filter_descriptors[5] = {
-        { MIP_DATA_DESC_SHARED_GPS_TIME,         filter_decimation },
+        { MIP_DATA_DESC_FILTER_FILTER_TIMESTAMP, filter_decimation },
         { MIP_DATA_DESC_FILTER_FILTER_STATUS,    filter_decimation },
         { MIP_DATA_DESC_FILTER_POS_LLH,          filter_decimation },
         { MIP_DATA_DESC_FILTER_VEL_NED,          filter_decimation },
         { MIP_DATA_DESC_FILTER_ATT_EULER_ANGLES, filter_decimation },
     };
 
-    if(mip_3dm_write_message_format(&device, MIP_FILTER_DATA_DESC_SET, 5, filter_descriptors) != MIP_ACK_OK)
+    if(mip_3dm_write_filter_message_format(&device, 5, filter_descriptors) != MIP_ACK_OK)
         exit_gracefully("ERROR: Could not set filter message format!");
 
 
     //
-    //Setup the sensor to vehicle transformation
+    //Setup the sensor to vehicle rotation
     //
 
-    if(mip_3dm_write_sensor_2_vehicle_transform_euler(&device, sensor_to_vehicle_transformation_euler[0], sensor_to_vehicle_transformation_euler[1], sensor_to_vehicle_transformation_euler[2]) != MIP_ACK_OK)
-        exit_gracefully("ERROR: Could not set sensor-to-vehicle transformation!");
+    if(mip_filter_write_sensor_to_vehicle_rotation_euler(&device, sensor_to_vehicle_rotation_euler[0], sensor_to_vehicle_rotation_euler[1], sensor_to_vehicle_rotation_euler[2]) != MIP_ACK_OK)
+        exit_gracefully("ERROR: Could not set sensor-2-vehicle rotation!");
+
 
     //
-    //Setup the GNSS antenna offsets
+    //Setup the GNSS antenna offset
     //
 
-    //GNSS1
-    if(mip_filter_write_multi_antenna_offset(&device, 1, gnss1_antenna_offset_meters) != MIP_ACK_OK)
+    if(mip_filter_write_antenna_offset(&device, gnss_antenna_offset_meters) != MIP_ACK_OK)
         exit_gracefully("ERROR: Could not set GNSS1 antenna offset!");
 
-    //GNSS2
-    if(mip_filter_write_multi_antenna_offset(&device, 2, gnss2_antenna_offset_meters) != MIP_ACK_OK)
-        exit_gracefully("ERROR: Could not set GNSS2 antenna offset!");
+
+    //
+    //Setup heading update control
+    //
+
+    if(mip_filter_write_heading_source(&device, MIP_FILTER_HEADING_SOURCE_COMMAND_SOURCE_GNSS_VEL_AND_MAG) != MIP_ACK_OK)
+        exit_gracefully("ERROR: Could not set filter heading update control!");
 
 
     //
-    //Setup the filter aiding measurements (GNSS position/velocity and dual antenna [aka gnss heading])
+    //Enable filter auto-initialization
     //
 
-    if(mip_filter_write_aiding_measurement_enable(&device, MIP_FILTER_AIDING_MEASUREMENT_ENABLE_COMMAND_AIDING_SOURCE_GNSS_POS_VEL, true) != MIP_ACK_OK)
-        exit_gracefully("ERROR: Could not set filter aiding measurement enable!");
-
-    if(mip_filter_write_aiding_measurement_enable(&device, MIP_FILTER_AIDING_MEASUREMENT_ENABLE_COMMAND_AIDING_SOURCE_GNSS_HEADING, true) != MIP_ACK_OK)
-        exit_gracefully("ERROR: Could not set filter aiding measurement enable!");
-
-
-    //
-    //Enable the wheeled-vehicle constraint
-    //
-
-    if(mip_filter_write_wheeled_vehicle_constraint_control(&device, 1) != MIP_ACK_OK)
-        exit_gracefully("ERROR: Could not set filter wheeled-vehicle constraint enable!");
-
-
-    //
-    //Setup the filter initialization (note: heading alignment is a bitfield!)
-    //
-
-    float filter_init_pos[3] = {0};
-    float filter_init_vel[3] = {0};
-
-    if(mip_filter_write_initialization_configuration(&device, 0, MIP_FILTER_INITIALIZATION_CONFIGURATION_COMMAND_INITIAL_CONDITION_SOURCE_AUTO_POS_VEL_ATT,
-       MIP_FILTER_INITIALIZATION_CONFIGURATION_COMMAND_ALIGNMENT_SELECTOR_DUAL_ANTENNA | MIP_FILTER_INITIALIZATION_CONFIGURATION_COMMAND_ALIGNMENT_SELECTOR_KINEMATIC,
-       0.0, 0.0, 0.0, filter_init_pos, filter_init_vel, MIP_FILTER_REFERENCE_FRAME_LLH) != MIP_ACK_OK)
-        exit_gracefully("ERROR: Could not set filter initialization configuration!");
+    if(mip_filter_write_auto_init_control(&device, 1) != MIP_ACK_OK)
+        exit_gracefully("ERROR: Could not set filter autoinit control!");
 
 
     //
@@ -278,21 +253,20 @@ int main(int argc, const char* argv[])
     //Sensor Data
     mip_dispatch_handler sensor_data_handlers[4];
 
-    mip_interface_register_extractor(&device, &sensor_data_handlers[0], MIP_SENSOR_DATA_DESC_SET, MIP_DATA_DESC_SHARED_GPS_TIME,     extract_mip_shared_gps_timestamp_data_from_field, &sensor_gps_time);
-    mip_interface_register_extractor(&device, &sensor_data_handlers[1], MIP_SENSOR_DATA_DESC_SET, MIP_DATA_DESC_SENSOR_ACCEL_SCALED, extract_mip_sensor_scaled_accel_data_from_field,  &sensor_accel);
-    mip_interface_register_extractor(&device, &sensor_data_handlers[2], MIP_SENSOR_DATA_DESC_SET, MIP_DATA_DESC_SENSOR_GYRO_SCALED,  extract_mip_sensor_scaled_gyro_data_from_field,   &sensor_gyro);
-    mip_interface_register_extractor(&device, &sensor_data_handlers[3], MIP_SENSOR_DATA_DESC_SET, MIP_DATA_DESC_SENSOR_MAG_SCALED,   extract_mip_sensor_scaled_mag_data_from_field,    &sensor_mag);
+    mip_interface_register_extractor(&device, &sensor_data_handlers[0], MIP_SENSOR_DATA_DESC_SET, MIP_DATA_DESC_SENSOR_TIME_STAMP_GPS, extract_mip_sensor_gps_timestamp_data_from_field,    &sensor_gps_time);
+    mip_interface_register_extractor(&device, &sensor_data_handlers[1], MIP_SENSOR_DATA_DESC_SET, MIP_DATA_DESC_SENSOR_ACCEL_SCALED,   extract_mip_sensor_scaled_accel_data_from_field, &sensor_accel);
+    mip_interface_register_extractor(&device, &sensor_data_handlers[2], MIP_SENSOR_DATA_DESC_SET, MIP_DATA_DESC_SENSOR_GYRO_SCALED,    extract_mip_sensor_scaled_gyro_data_from_field,  &sensor_gyro);
+    mip_interface_register_extractor(&device, &sensor_data_handlers[3], MIP_SENSOR_DATA_DESC_SET, MIP_DATA_DESC_SENSOR_MAG_SCALED,     extract_mip_sensor_scaled_mag_data_from_field,   &sensor_mag);
 
     //GNSS Data
-    mip_dispatch_handler gnss_data_handlers[2];
+    mip_dispatch_handler gnss_data_handlers[1];
 
-    mip_interface_register_extractor(&device, &gnss_data_handlers[0], MIP_GNSS1_DATA_DESC_SET, MIP_DATA_DESC_GNSS_FIX_INFO, extract_mip_gnss_fix_info_data_from_field, &gnss_fix_info[0]);
-    mip_interface_register_extractor(&device, &gnss_data_handlers[1], MIP_GNSS2_DATA_DESC_SET, MIP_DATA_DESC_GNSS_FIX_INFO, extract_mip_gnss_fix_info_data_from_field, &gnss_fix_info[1]);
+    mip_interface_register_extractor(&device, &gnss_data_handlers[0], MIP_GNSS1_DATA_DESC_SET, MIP_DATA_DESC_GNSS_FIX_INFO, extract_mip_gnss_fix_info_data_from_field, &gnss_fix_info);
 
     //Filter Data
     mip_dispatch_handler filter_data_handlers[5];
 
-    mip_interface_register_extractor(&device, &filter_data_handlers[0], MIP_FILTER_DATA_DESC_SET, MIP_DATA_DESC_SHARED_GPS_TIME,         extract_mip_shared_gps_timestamp_data_from_field, &filter_gps_time);
+    mip_interface_register_extractor(&device, &filter_data_handlers[0], MIP_FILTER_DATA_DESC_SET, MIP_DATA_DESC_FILTER_FILTER_TIMESTAMP, extract_mip_filter_timestamp_data_from_field,     &filter_gps_time);
     mip_interface_register_extractor(&device, &filter_data_handlers[1], MIP_FILTER_DATA_DESC_SET, MIP_DATA_DESC_FILTER_FILTER_STATUS,    extract_mip_filter_status_data_from_field,        &filter_status);
     mip_interface_register_extractor(&device, &filter_data_handlers[2], MIP_FILTER_DATA_DESC_SET, MIP_DATA_DESC_FILTER_POS_LLH,          extract_mip_filter_position_llh_data_from_field,  &filter_position_llh);
     mip_interface_register_extractor(&device, &filter_data_handlers[3], MIP_FILTER_DATA_DESC_SET, MIP_DATA_DESC_FILTER_VEL_NED,          extract_mip_filter_velocity_ned_data_from_field,  &filter_velocity_ned);
@@ -313,37 +287,30 @@ int main(int argc, const char* argv[])
     bool          running              = true;
     mip_timestamp prev_print_timestamp = 0;
 
-    printf("Sensor is configured... waiting for filter to enter Full Navigation mode.\n");
+    printf("Sensor is configured... waiting for filter to enter running mode.\n");
 
-    char *state_init = "";
-    char **current_state = &state_init;
     while(running)
     {
         mip_interface_update(&device, 0, false);
-        displayFilterState(filter_status.filter_state, current_state, false);
+
 
         //Check GNSS fixes and alert the user when they become valid
-        for(int i=0; i<2; i++)
+        if((gnss_fix_info_valid == false) && (gnss_fix_info.fix_type == MIP_GNSS_FIX_INFO_DATA_FIX_TYPE_FIX_3D) &&
+            (gnss_fix_info.valid_flags & MIP_GNSS_FIX_INFO_DATA_VALID_FLAGS_FIX_TYPE))
         {
-            if((gnss_fix_info_valid[i] == false) && ((gnss_fix_info[i].fix_type == MIP_GNSS_FIX_INFO_DATA_FIX_TYPE_FIX_3D) ||
-                                                     (gnss_fix_info[i].fix_type == MIP_GNSS_FIX_INFO_DATA_FIX_TYPE_FIX_RTK_FLOAT) ||
-                                                     (gnss_fix_info[i].fix_type == MIP_GNSS_FIX_INFO_DATA_FIX_TYPE_FIX_RTK_FIXED)) &&
-                (gnss_fix_info[i].valid_flags & MIP_GNSS_FIX_INFO_DATA_VALID_FLAGS_FIX_TYPE))
-            {
-                printf("NOTE: GNSS%i fix info valid.\n", i+1);
-                gnss_fix_info_valid[i] = true;
-            }
+            printf("NOTE: GNSS fix info valid.\n");
+            gnss_fix_info_valid = true;
         }
 
         //Check Filter State
-        if((!filter_state_full_nav) && (filter_status.filter_state == MIP_FILTER_MODE_FULL_NAV))
+        if((!filter_state_running) && ((filter_status.filter_state == MIP_FILTER_MODE_GX5_RUN_SOLUTION_ERROR) || (filter_status.filter_state == MIP_FILTER_MODE_GX5_RUN_SOLUTION_VALID)))
         {
-            printf("NOTE: Filter has entered full navigation mode.\n");
-            filter_state_full_nav = true;
+            printf("NOTE: Filter has entered running mode.\n");
+            filter_state_running = true;
         }
 
-        //Once in full nav, print out data at 1 Hz
-        if(filter_state_full_nav)
+        //Once in running mode, print out data at 1 Hz
+        if(filter_state_running)
         {
            mip_timestamp curr_time = get_current_timestamp();
 
@@ -363,6 +330,7 @@ int main(int argc, const char* argv[])
 
     exit_gracefully("Example Completed Successfully.");
 }
+
 
 ////////////////////////////////////////////////////////////////////////////////
 // Print Usage Function
