+++ examples/c/stream_imu/stream_imu.c
@@ -1,14 +1,14 @@
 
-#include "example_utils.h"
+#include "../example_utils.h"
 
-#include <mip/mip_dispatch.h>
-#include <mip/mip_interface.h>
-#include <microstrain/connections/serial/serial_port.h>
-#include <microstrain/common/platform.h>
+#include <../../src/c/mip/mip_dispatch.h>
+#include <../../src/c/mip/mip_interface.h>
+#include <../../src/c/microstrain/connections/serial/serial_port.h>
+#include <../../src/c/microstrain/common/platform.h>
 
-#include <mip/definitions/commands_base.h>
-#include <mip/definitions/commands_3dm.h>
-#include <mip/definitions/data_sensor.h>
+#include <../../src/c/mip/definitions/commands_base.h>
+#include <../../src/c/mip/definitions/commands_3dm.h>
+#include <../../src/c/mip/definitions/data_sensor.h>
 
 #include <stdio.h>
 #include <stdlib.h>
