+++ examples/cpp/cv7/cv7_example.cpp
@@ -1,14 +1,14 @@
 
 /////////////////////////////////////////////////////////////////////////////
 //
-// GQ7_Example.cpp
+// CV7_Example.cpp
 //
-// C++ Example set-up program for the GQ7
+// C++ Example set-up program for the CV7
 //
-// This example shows a typical setup for the GQ7 sensor in a wheeled-vehicle application using C++.
-// It is not an exhaustive example of all GQ7 settings.
+// This example shows a typical setup for the CV7 sensor using C++.
+// It is not an exhaustive example of all CV7 settings.
 // If your specific setup needs are not met by this example, please consult
-// the MSCL-embedded API documentation for the proper commands.
+// the MIP SDK API documentation for the proper commands.
 //
 //
 //!@section LICENSE
@@ -27,12 +27,11 @@
 // Include Files
 ////////////////////////////////////////////////////////////////////////////////
 
-#include <mip/mip_all.hpp>
+#include "../example_utils.hpp"
+
+#include <../../src/cpp/mip/mip_all.hpp>
 
 #include <array>
-
-#include "example_utils.hpp"
-
 
 ////////////////////////////////////////////////////////////////////////////////
 // Global Variables
@@ -41,27 +40,21 @@
 //Sensor-to-vehicle frame transformation (Euler Angles)
 float sensor_to_vehicle_transformation_euler[3] = {0.0, 0.0, 0.0};
 
-//GNSS antenna offsets
-float gnss1_antenna_offset_meters[3] = {-0.25, 0.0, 0.0};
-float gnss2_antenna_offset_meters[3] = {0.25, 0.0, 0.0};
-
 //Device data stores
 mip::data_shared::GpsTimestamp sensor_gps_time;
 mip::data_sensor::ScaledAccel  sensor_accel;
 mip::data_sensor::ScaledGyro   sensor_gyro;
 mip::data_sensor::ScaledMag    sensor_mag;
 
-mip::data_gnss::FixInfo        gnss_fix_info[2];
-
-bool gnss_fix_info_valid[2] = {false};
-
 mip::data_shared::GpsTimestamp filter_gps_time;
 mip::data_filter::Status       filter_status;
-mip::data_filter::PositionLlh  filter_position_llh;
-mip::data_filter::VelocityNed  filter_velocity_ned;
 mip::data_filter::EulerAngles  filter_euler_angles;
 
-bool filter_state_full_nav = false;
+bool filter_state_ahrs = false;
+
+
+const uint8_t FILTER_ROLL_EVENT_ACTION_ID  = 1;
+const uint8_t FILTER_PITCH_EVENT_ACTION_ID = 2;
 
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -73,6 +66,7 @@ int usage(const char* argv0);
 void exit_gracefully(const char *message);
 bool should_exit();
 
+void handleFilterEventSource(void*, const mip::FieldView& field, mip::Timestamp timestamp);
 
 ////////////////////////////////////////////////////////////////////////////////
 // Main Function
@@ -93,7 +87,6 @@ int main(int argc, const char* argv[])
     }
 
     std::unique_ptr<mip::Interface>& device = utils->device;
-    printf("Connecting to and configuring sensor.\n");
 
     //
     //Ping the device (note: this is good to do to make sure the device is present)
@@ -146,23 +139,6 @@ int main(int argc, const char* argv[])
 
 
     //
-    //Setup GNSS 1 and 2 data format to 2 Hz (decimation of 1)
-    //
-
-    std::array<mip::DescriptorRate, 1> gnss_descriptors = {{
-        { mip::data_gnss::DATA_FIX_INFO, 1 }
-    }};
-
-    //GNSS1
-    if(mip::commands_3dm::writeMessageFormat(*device, mip::data_gnss::MIP_GNSS1_DATA_DESC_SET, static_cast<uint8_t>(gnss_descriptors.size()), gnss_descriptors.data()) != mip::CmdResult::ACK_OK)
-        exit_gracefully("ERROR: Could not set GNSS1 message format!");
-
-    //GNSS2
-    if(mip::commands_3dm::writeMessageFormat(*device, mip::data_gnss::MIP_GNSS2_DATA_DESC_SET, static_cast<uint8_t>(gnss_descriptors.size()), gnss_descriptors.data()) != mip::CmdResult::ACK_OK)
-        exit_gracefully("ERROR: Could not set GNSS2 message format!");
-
-
-    //
     //Setup FILTER data format
     //
 
@@ -174,17 +150,68 @@ int main(int argc, const char* argv[])
     const uint16_t filter_sample_rate = 100; // Hz
     const uint16_t filter_decimation = filter_base_rate / filter_sample_rate;
 
-    std::array<mip::DescriptorRate, 5> filter_descriptors = {{
+    std::array<mip::DescriptorRate, 3> filter_descriptors = {{
         { mip::data_shared::DATA_GPS_TIME,         filter_decimation },
         { mip::data_filter::DATA_FILTER_STATUS,    filter_decimation },
-        { mip::data_filter::DATA_POS_LLH,          filter_decimation },
-        { mip::data_filter::DATA_VEL_NED,          filter_decimation },
         { mip::data_filter::DATA_ATT_EULER_ANGLES, filter_decimation },
     }};
 
     if(mip::commands_3dm::writeMessageFormat(*device, mip::data_filter::DESCRIPTOR_SET, static_cast<uint8_t>(filter_descriptors.size()), filter_descriptors.data()) != mip::CmdResult::ACK_OK)
         exit_gracefully("ERROR: Could not set filter message format!");
 
+
+    //
+    // Setup event triggers/actions on > 45 degrees filter pitch and roll Euler angles
+    // (Note 1: we are reusing the event and action structs, since the settings for pitch/roll are so similar)
+    // (Note 2: we are using the same value for event and action ids.  This is not necessary, but done here for convenience)
+    //
+
+    //EVENTS
+
+    //Roll
+    mip::commands_3dm::EventTrigger::Parameters event_params;
+    event_params.threshold.type       = mip::commands_3dm::EventTrigger::ThresholdParams::Type::WINDOW;
+    event_params.threshold.desc_set   = mip::data_filter::DESCRIPTOR_SET;
+    event_params.threshold.field_desc = mip::data_filter::DATA_ATT_EULER_ANGLES;
+    event_params.threshold.param_id   = 1;
+    event_params.threshold.high_thres = -0.7853981;
+    event_params.threshold.low_thres  = 0.7853981;
+
+    if(mip::commands_3dm::writeEventTrigger(*device, FILTER_ROLL_EVENT_ACTION_ID, mip::commands_3dm::EventTrigger::Type::THRESHOLD, event_params) != mip::CmdResult::ACK_OK)
+        exit_gracefully("ERROR: Could not set roll event parameters!");
+
+    //Pitch
+    event_params.threshold.param_id = 2;
+
+    if(mip::commands_3dm::writeEventTrigger(*device, FILTER_PITCH_EVENT_ACTION_ID, mip::commands_3dm::EventTrigger::Type::THRESHOLD, event_params) != mip::CmdResult::ACK_OK)
+        exit_gracefully("ERROR: Could not set pitch event parameters!");
+
+    //ACTIONS
+
+    //Roll
+    mip::commands_3dm::EventAction::Parameters event_action;
+    event_action.message.desc_set       = mip::data_filter::DESCRIPTOR_SET;
+    event_action.message.num_fields     = 1;
+    event_action.message.descriptors[0] = mip::data_shared::DATA_EVENT_SOURCE;
+    event_action.message.decimation     = 0;
+
+    if(writeEventAction(*device, FILTER_ROLL_EVENT_ACTION_ID, FILTER_ROLL_EVENT_ACTION_ID, mip::commands_3dm::EventAction::Type::MESSAGE, event_action) != mip::CmdResult::ACK_OK)
+        exit_gracefully("ERROR: Could not set roll action parameters!");
+
+    //Pitch
+    if(writeEventAction(*device, FILTER_PITCH_EVENT_ACTION_ID, FILTER_PITCH_EVENT_ACTION_ID, mip::commands_3dm::EventAction::Type::MESSAGE, event_action) != mip::CmdResult::ACK_OK)
+        exit_gracefully("ERROR: Could not set pitch action parameters!");
+
+
+    //ENABLE EVENTS
+
+    //Roll
+    if(writeEventControl(*device, FILTER_ROLL_EVENT_ACTION_ID, mip::commands_3dm::EventControl::Mode::ENABLED) != mip::CmdResult::ACK_OK)
+        exit_gracefully("ERROR: Could not enable roll event!");
+
+    //Pitch
+    if(writeEventControl(*device, FILTER_PITCH_EVENT_ACTION_ID, mip::commands_3dm::EventControl::Mode::ENABLED) != mip::CmdResult::ACK_OK)
+        exit_gracefully("ERROR: Could not enable pitch event!");
 
     //
     //Setup the sensor to vehicle transformation
@@ -195,50 +222,11 @@ int main(int argc, const char* argv[])
 
 
     //
-    //Setup the GNSS antenna offsets
-    //
-
-    //GNSS1
-    if(mip::commands_filter::writeMultiAntennaOffset(*device, 1, gnss1_antenna_offset_meters) != mip::CmdResult::ACK_OK)
-        exit_gracefully("ERROR: Could not set GNSS1 antenna offset!");
-
-    //GNSS2
-    if(mip::commands_filter::writeMultiAntennaOffset(*device, 2, gnss2_antenna_offset_meters) != mip::CmdResult::ACK_OK)
-        exit_gracefully("ERROR: Could not set GNSS2 antenna offset!");
-
-
-    //
     //Setup the filter aiding measurements (GNSS position/velocity and dual antenna [aka gnss heading])
     //
 
-    if(mip::commands_filter::writeAidingMeasurementEnable(*device, mip::commands_filter::AidingMeasurementEnable::AidingSource::GNSS_POS_VEL, true) != mip::CmdResult::ACK_OK)
+    if(mip::commands_filter::writeAidingMeasurementEnable(*device, mip::commands_filter::AidingMeasurementEnable::AidingSource::MAGNETOMETER, true) != mip::CmdResult::ACK_OK)
         exit_gracefully("ERROR: Could not set filter aiding measurement enable!");
-
-    if(mip::commands_filter::writeAidingMeasurementEnable(*device, mip::commands_filter::AidingMeasurementEnable::AidingSource::GNSS_HEADING, true) != mip::CmdResult::ACK_OK)
-        exit_gracefully("ERROR: Could not set filter aiding measurement enable!");
-
-
-    //
-    //Enable the wheeled-vehicle constraint
-    //
-
-    if(mip::commands_filter::writeWheeledVehicleConstraintControl(*device, 1) != mip::CmdResult::ACK_OK)
-        exit_gracefully("ERROR: Could not set filter wheeled-vehicle constraint enable!");
-
-
-    //
-    //Setup the filter initialization (note: heading alignment is a bitfield!)
-    //
-
-    float filter_init_pos[3] = {0};
-    float filter_init_vel[3] = {0};
-
-    mip::commands_filter::InitializationConfiguration::AlignmentSelector alignment;
-    alignment = alignment.DUAL_ANTENNA | alignment.KINEMATIC;
-
-    if(mip::commands_filter::writeInitializationConfiguration(*device, 0, mip::commands_filter::InitializationConfiguration::InitialConditionSource::AUTO_POS_VEL_ATT,
-       alignment, 0.0, 0.0, 0.0, filter_init_pos, filter_init_vel, mip::commands_filter::FilterReferenceFrame::LLH) != mip::CmdResult::ACK_OK)
-        exit_gracefully("ERROR: Could not set filter initialization configuration!");
 
 
     //
@@ -261,23 +249,14 @@ int main(int argc, const char* argv[])
     device->registerExtractor(sensor_data_handlers[2], &sensor_gyro);
     device->registerExtractor(sensor_data_handlers[3], &sensor_mag);
 
-
-    //GNSS Data
-    mip::DispatchHandler gnss_data_handlers[2];
-
-    device->registerExtractor(gnss_data_handlers[0], &gnss_fix_info[0], mip::data_gnss::MIP_GNSS1_DATA_DESC_SET);
-    device->registerExtractor(gnss_data_handlers[1], &gnss_fix_info[1], mip::data_gnss::MIP_GNSS2_DATA_DESC_SET);
-
-
     //Filter Data
-    mip::DispatchHandler filter_data_handlers[5];
+    mip::DispatchHandler filter_data_handlers[4];
 
     device->registerExtractor(filter_data_handlers[0], &filter_gps_time, mip::data_filter::DESCRIPTOR_SET);
     device->registerExtractor(filter_data_handlers[1], &filter_status);
-    device->registerExtractor(filter_data_handlers[2], &filter_position_llh);
-    device->registerExtractor(filter_data_handlers[3], &filter_velocity_ned);
-    device->registerExtractor(filter_data_handlers[4], &filter_euler_angles);
+    device->registerExtractor(filter_data_handlers[2], &filter_euler_angles);
 
+    device->registerFieldCallback<&handleFilterEventSource>(filter_data_handlers[3], mip::data_filter::DESCRIPTOR_SET, mip::data_shared::DATA_EVENT_SOURCE);
 
     //
     //Resume the device
@@ -294,44 +273,30 @@ int main(int argc, const char* argv[])
     bool running = true;
     mip::Timestamp prev_print_timestamp = getCurrentTimestamp();
 
-    printf("Sensor is configured... waiting for filter to enter Full Navigation mode (FULL_NAV).\n");
+    printf("Sensor is configured... waiting for filter to enter AHRS mode (AHRS).\n");
 
-    std::string current_state = std::string{""};
-    while(running) {
+    std::string current_state = "";
+    while(running)
+    {
         device->update();
         displayFilterState(filter_status.filter_state, current_state);
 
-        //Check GNSS fixes and alert the user when they become valid
-        for(int i=0; i<2; i++)
-        {
-            if((gnss_fix_info_valid[i] == false) && ((gnss_fix_info[i].fix_type == mip::data_gnss::FixInfo::FixType::FIX_3D) ||
-                                                     (gnss_fix_info[i].fix_type == mip::data_gnss::FixInfo::FixType::FIX_RTK_FLOAT) ||
-                                                     (gnss_fix_info[i].fix_type == mip::data_gnss::FixInfo::FixType::FIX_RTK_FIXED)) &&
-                (gnss_fix_info[i].valid_flags & mip::data_gnss::FixInfo::ValidFlags::FIX_TYPE))
-            {
-                printf("NOTE: GNSS%i fix info valid.\n", i+1);
-                gnss_fix_info_valid[i] = true;
-            }
-        }
-
         //Check Filter State
-        if((!filter_state_full_nav) && (filter_status.filter_state == mip::data_filter::FilterMode::FULL_NAV))
+        if((!filter_state_ahrs) && (filter_status.filter_state == mip::data_filter::FilterMode::AHRS))
         {
-            printf("NOTE: Filter has entered full navigation mode.\n");
-            filter_state_full_nav = true;
+            printf("NOTE: Filter has entered AHRS mode.\n");
+            filter_state_ahrs = true;
         }
 
-        //Once in full nav, print out data at 1 Hz
-        if(filter_state_full_nav)
+        //Once in AHRS Flter Mode, print out data at 10 Hz
+        if(filter_state_ahrs)
         {
            mip::Timestamp curr_timestamp = getCurrentTimestamp();
 
            if(curr_timestamp - prev_print_timestamp >= 1000)
            {
-                printf("TOW = %f: POS_LLH = [%f, %f, %f], VEL_NED = [%f, %f, %f], ATT_EULER = [%f %f %f]\n",
-                       filter_gps_time.tow, filter_position_llh.latitude, filter_position_llh.longitude, filter_position_llh.ellipsoid_height,
-                       filter_velocity_ned.north, filter_velocity_ned.east, filter_velocity_ned.down,
-                       filter_euler_angles.roll, filter_euler_angles.pitch, filter_euler_angles.yaw);
+                printf("TOW = %f: ATT_EULER = [%f %f %f]\n",
+                       filter_gps_time.tow, filter_euler_angles.roll, filter_euler_angles.pitch, filter_euler_angles.yaw);
 
                 prev_print_timestamp = curr_timestamp;
            }
@@ -340,7 +305,26 @@ int main(int argc, const char* argv[])
         running = !should_exit();
     }
 
+
     exit_gracefully("Example Completed Successfully.");
+}
+
+
+////////////////////////////////////////////////////////////////////////////////
+// Filter Event Source Field Handler
+////////////////////////////////////////////////////////////////////////////////
+
+void handleFilterEventSource(void*, const mip::FieldView& field, mip::Timestamp)
+{
+    mip::data_shared::EventSource data;
+
+    if(field.extract(data))
+    {
+      if(data.trigger_id == FILTER_ROLL_EVENT_ACTION_ID)
+        printf("WARNING: Roll event triggered!\n");
+      else if(data.trigger_id == FILTER_PITCH_EVENT_ACTION_ID)
+        printf("WARNING: Pitch event triggered!\n");
+    }
 }
