+++ examples/cpp/cx5_gx5_cv5_15_25/cx5_gx5_cv5_15_25_example.cpp
@@ -8,7 +8,7 @@
 // This example shows a typical setup for the CX5-15 sensor in a wheeled-vehicle application using using C++.
 // It is not an exhaustive example of all CX5-15 settings.
 // If your specific setup needs are not met by this example, please consult
-// the MSCL-embedded API documentation for the proper commands.
+// the MIP SDK API documentation for the proper commands.
 //
 //
 //!@section LICENSE
@@ -27,7 +27,7 @@
 // Include Files
 ////////////////////////////////////////////////////////////////////////////////
 
-#include <mip/mip_all.hpp>
+#include <../../src/cpp/mip/mip_all.hpp>
 #include <array>
 #include "../example_utils.hpp"
