+++ examples/cpp/pretty_printer/pretty_printer.cpp
