+++ examples/cpp/pretty_printer/stringify.cpp
