+++ examples/cpp/pretty_printer/stringify.hpp
