+++ examples/cpp/stream_imu/stream_imu.cpp
@@ -1,9 +1,9 @@
 
-#include "example_utils.hpp"
+#include "../example_utils.hpp"
 
-#include <mip/definitions/commands_base.hpp>
-#include <mip/definitions/commands_3dm.hpp>
-#include <mip/definitions/data_sensor.hpp>
+#include <../../src/cpp/mip/definitions/commands_base.hpp>
+#include <../../src/cpp/mip/definitions/commands_3dm.hpp>
+#include <../../src/cpp/mip/definitions/data_sensor.hpp>
 
 #include <stdexcept>
 #include <thread>
